+++ graphql-gradle-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/GraphQLVariablesIT.java
@@ -10,6 +10,7 @@ import java.util.GregorianCalendar;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 import org.allGraphQLCases.client.AnotherMutationType;
 import org.allGraphQLCases.client.CharacterInput;
@@ -102,8 +103,8 @@ class GraphQLVariablesIT {
 				+ " $Value :   String ! , $anotherValue:String) {directiveOnQuery (uppercase: $uppercase) @testDirective(value:$Value, anotherValue:$anotherValue)}");
 		Map<String, Object> params = new HashMap<>();
 		params.put("uppercase", true);
-		params.put("anotherValue", "another value");
-		params.put("Value", "a first value");
+		params.put("anotherValue", "another value with an antislash: \\");
+		params.put("Value", "a first \"value\"");
 
 		// Go, go, go
 		MyQueryType resp = directiveOnQuery.execQuery(params);
@@ -114,8 +115,8 @@ class GraphQLVariablesIT {
 		assertNotNull(ret);
 		assertEquals(2, ret.size());
 		//
-		assertEquals("A FIRST VALUE", ret.get(0));
-		assertEquals("ANOTHER VALUE", ret.get(1));
+		assertEquals("A FIRST \"VALUE\"", ret.get(0));
+		assertEquals("ANOTHER VALUE WITH AN ANTISLASH: \\", ret.get(1));
 	}
 
 	@Execution(ExecutionMode.CONCURRENT)
@@ -196,20 +197,14 @@ class GraphQLVariablesIT {
 
 		// Let's wait a max of 10 second, until we receive some notifications (my PC is really slow, especially when the
 		// antivirus consumes 98% of my CPU!
-		try {
-			for (int i = 1; i < 100; i += 1) {
-				if (callback.lastReceivedMessage != null)
-					break;
-				Thread.sleep(100); // Wait 0.1 second
-			} // for
-		} catch (InterruptedException ex) {
-			Thread.currentThread().interrupt();
-		}
+		callback.latchForMessageReception.await(10, TimeUnit.SECONDS);
 
 		// Let's disconnect from the subscription
 		sub.unsubscribe();
 
 		// Verification
+		assertNull(callback.lastReceivedError, "expected no error, but received " + callback.lastReceivedError);
 		assertNotNull(callback.lastReceivedMessage, "The subscription should have received a message");
 	}
+
 }