+++ graphql-gradle-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/SpringTestConfig.java
@@ -4,17 +4,23 @@ import org.allGraphQLCases.client.util.MyQueryTypeExecutorAllGraphQLCases;
 import org.allGraphQLCases.client2.util.MyQueryTypeExecutorAllGraphQLCases2;
 import org.allGraphQLCases.demo.impl.PartialDirectQueries;
 import org.forum.client.util.QueryExecutorForum;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.security.oauth2.client.registration.ReactiveClientRegistrationRepository;
 import org.springframework.security.oauth2.client.web.reactive.function.client.ServerOAuth2AuthorizedClientExchangeFilterFunction;
 import org.springframework.security.oauth2.client.web.server.UnAuthenticatedServerOAuth2AuthorizedClientRepository;
+import org.springframework.web.reactive.function.client.WebClient;
 
 import com.graphql_java_generator.client.GraphQLConfiguration;
 import com.graphql_java_generator.client.graphqlrepository.EnableGraphQLRepositories;
+
+import reactor.netty.http.client.HttpClient;
 
 @Configuration
 @SpringBootApplication
@@ -27,6 +33,7 @@ import com.graphql_java_generator.client.graphqlrepository.EnableGraphQLReposito
 public class SpringTestConfig {
 
 	@Bean
+	@Primary
 	ServerOAuth2AuthorizedClientExchangeFilterFunction serverOAuth2AuthorizedClientExchangeFilterFunctionAllGraphQLCases(
 			ReactiveClientRegistrationRepository clientRegistrations) {
 		ServerOAuth2AuthorizedClientExchangeFilterFunction oauth = new ServerOAuth2AuthorizedClientExchangeFilterFunction(
@@ -36,8 +43,18 @@ public class SpringTestConfig {
 	}
 
 	@Bean
+	@Primary
 	ServerOAuth2AuthorizedClientExchangeFilterFunction serverOAuth2AuthorizedClientExchangeFilterFunctionForum(
 			ReactiveClientRegistrationRepository clientRegistrations) {
 		return null;
+	}
+
+	@Bean
+	@Primary
+	public WebClient webClientAllGraphQLCases(String graphqlEndpointAllGraphQLCases, //
+			@Autowired(required = false) @Qualifier("httpClientAllGraphQLCases") HttpClient httpClientAllGraphQLCases,
+			@Autowired(required = false) @Qualifier("serverOAuth2AuthorizedClientExchangeFilterFunctionAllGraphQLCases") ServerOAuth2AuthorizedClientExchangeFilterFunction serverOAuth2AuthorizedClientExchangeFilterFunctionAllGraphQLCases) {
+		return GraphQLConfiguration.getWebClient(graphqlEndpointAllGraphQLCases, httpClientAllGraphQLCases,
+				serverOAuth2AuthorizedClientExchangeFilterFunctionAllGraphQLCases);
 	}
 }