+++ graphql-gradle-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/subscription/SubscriptionCallbackListInteger.java
@@ -4,6 +4,7 @@
 package org.allGraphQLCases.subscription;
 
 import java.util.List;
+import java.util.concurrent.CountDownLatch;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,15 +16,18 @@ import com.graphql_java_generator.client.SubscriptionCallback;
  * 
  * @author etienne-sf
  */
-public class SubscriptionCallbackListIntegerForTest implements SubscriptionCallback<List<Integer>> {
+public class SubscriptionCallbackListInteger implements SubscriptionCallback<List<Integer>> {
 
 	/** The logger for this class */
-	static protected Logger logger = LoggerFactory.getLogger(SubscriptionCallbackListIntegerForTest.class);
+	static protected Logger logger = LoggerFactory.getLogger(SubscriptionCallbackListInteger.class);
 
 	final String clientName;
 	public List<Integer> lastReceivedMessage = null;
 
-	public SubscriptionCallbackListIntegerForTest(String clientName) {
+	/** A latch that will be freed when a the first notification arrives for this subscription */
+	public CountDownLatch latchForMessageReception = new CountDownLatch(1);
+
+	public SubscriptionCallbackListInteger(String clientName) {
 		this.clientName = clientName;
 	}
 
@@ -36,6 +40,7 @@ public class SubscriptionCallbackListIntegerForTest implements SubscriptionCallb
 	public void onMessage(List<Integer> t) {
 		logger.debug("Received this list from the 'subscribeToAList' subscription: {} (for {})", t, clientName);
 		lastReceivedMessage = t;
+		latchForMessageReception.countDown();
 	}
 
 	@Override
@@ -45,7 +50,8 @@ public class SubscriptionCallbackListIntegerForTest implements SubscriptionCallb
 
 	@Override
 	public void onError(Throwable cause) {
-		logger.error("Oups! An error occurred: " + cause.getMessage());
+		logger.error("Oups! An error occurred: "
+				+ ((cause == null) ? null : cause.getClass().getSimpleName() + ": " + cause.getMessage()));
 	}
 
 }