+++ graphql-gradle-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/two_graphql_servers/RequestsAgainstTwoGraphQLServersIT.java
@@ -11,11 +11,12 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.Calendar;
 import java.util.GregorianCalendar;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 import org.allGraphQLCases.SpringTestConfig;
 import org.allGraphQLCases.client.Character;
 import org.allGraphQLCases.client.util.MyQueryTypeExecutorAllGraphQLCases;
-import org.allGraphQLCases.subscription.SubscriptionCallbackListIntegerForTest;
+import org.allGraphQLCases.subscription.SubscriptionCallbackListInteger;
 import org.forum.client.Board;
 import org.forum.client.Member;
 import org.forum.client.MemberType;
@@ -97,31 +98,17 @@ public class RequestsAgainstTwoGraphQLServersIT {
 
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	void test_GraphQLRepository_allGraphQLCases() throws GraphQLRequestExecutionException {
+	void test_GraphQLRepository_allGraphQLCases() throws GraphQLRequestExecutionException, InterruptedException {
 		// Preparation
-		SubscriptionCallbackListIntegerForTest callback = new SubscriptionCallbackListIntegerForTest(
+		SubscriptionCallbackListInteger callback = new SubscriptionCallbackListInteger(
 				"FullRequestSubscriptionIT.test_SubscribeToAList");
 
 		// Go, go, go
 		SubscriptionClient sub = graphQLRepoAllGraphQLCases.subscribeToAList(callback);
 
 		// Verification
-		try {
-			Thread.sleep(500); // Wait 0.5 second, so that other thread is ready
-		} catch (InterruptedException ex) {
-			Thread.currentThread().interrupt();
-		}
-
-		// Let's wait a max of 10 second, until we receive some notifications
-		try {
-			for (int i = 1; i < 100; i += 1) {
-				if (callback.lastReceivedMessage != null)
-					break;
-				Thread.sleep(100); // Wait 0.1 second
-			} // for
-		} catch (InterruptedException ex) {
-			Thread.currentThread().interrupt();
-		}
+		// Let's wait a max of 20 second, until we receive some notifications
+		callback.latchForMessageReception.await(20, TimeUnit.SECONDS);
 
 		// Let's disconnect from the subscription
 		sub.unsubscribe();